# feat!: use crypto.getRandomValues instead of randomFillSync