This improves compatibility with other server runtimes such as Cloudflare Workers that don't support Node's crypto module out of the box.

I tried to dig into the history of the use of `randomFillSync`. The change was done in #435 but a justification was not mentioned in the commit or the PR.

If this is approved it would also open up the possibility of merging the two RNG implementations for node and browsers into a common one.

Benchmark before:

```
uuid.stringify() x 5,294,261 ops/sec ±2.28% (90 runs sampled)
uuid.parse() x 4,008,788 ops/sec ±0.90% (96 runs sampled)
---

uuid.v1() x 4,394,917 ops/sec ±0.75% (96 runs sampled)
uuid.v1() fill existing array x 7,200,421 ops/sec ±0.69% (93 runs sampled)
uuid.v4() x 10,264,478 ops/sec ±0.60% (97 runs sampled)
uuid.v4() fill existing array x 9,653,512 ops/sec ±0.70% (95 runs sampled)
uuid.v4() without native generation x 4,402,382 ops/sec ±0.34% (97 runs sampled)
uuid.v3() x 748,220 ops/sec ±0.29% (98 runs sampled)
uuid.v5() x 791,018 ops/sec ±0.26% (97 runs sampled)
uuid.v6() x 2,472,765 ops/sec ±0.25% (91 runs sampled)
uuid.v7() x 4,784,766 ops/sec ±0.47% (98 runs sampled)
uuid.v7() fill existing array x 8,528,953 ops/sec ±0.34% (98 runs sampled)
uuid.v7() with defined time x 5,180,108 ops/sec ±0.26% (100 runs sampled)
Fastest is uuid.v4()
---

uuid.v1ToV6() x 2,500,811 ops/sec ±0.42% (97 runs sampled)
uuid.v6ToV1() x 2,448,658 ops/sec ±0.72% (94 runs sampled)
```

Benchmark after:

```
uuid.stringify() x 5,287,311 ops/sec ±0.48% (94 runs sampled)
uuid.parse() x 4,070,022 ops/sec ±0.29% (96 runs sampled)
---

uuid.v1() x 3,562,691 ops/sec ±0.17% (98 runs sampled)
uuid.v1() fill existing array x 7,241,696 ops/sec ±0.20% (99 runs sampled)
uuid.v4() x 10,376,376 ops/sec ±0.31% (96 runs sampled)
uuid.v4() fill existing array x 10,066,924 ops/sec ±0.30% (99 runs sampled)
uuid.v4() without native generation x 4,442,705 ops/sec ±0.46% (98 runs sampled)
uuid.v3() x 687,526 ops/sec ±3.20% (92 runs sampled)
uuid.v5() x 757,411 ops/sec ±0.99% (92 runs sampled)
uuid.v6() x 2,428,337 ops/sec ±0.30% (95 runs sampled)
uuid.v7() x 4,740,402 ops/sec ±0.73% (95 runs sampled)
uuid.v7() fill existing array x 8,496,746 ops/sec ±0.54% (96 runs sampled)
uuid.v7() with defined time x 5,193,370 ops/sec ±0.23% (99 runs sampled)
Fastest is uuid.v4()
---

uuid.v1ToV6() x 2,503,294 ops/sec ±0.17% (97 runs sampled)
uuid.v6ToV1() x 2,500,651 ops/sec ±0.21% (98 runs sampled)
```