@broofa 2025-08-03
> The change was done in #435 but a justification was not mentioned in the commit or the PR.

@awwit might recall better, but I believe the switch to `randomFillSync`  was made because it allows for the reuse of the `rnds8Pool` array, which provides some performance benefits.