@realityking 2025-08-03
> @awwit might recall better, but I believe the switch to randomFillSync was made because it allows for the reuse of the rnds8Pool array, which provides some performance benefits.

The pool was introduced later in #513. #435 was about reusing the same `Uint8Array`. But I had indeed misread, it was a switch from `randomBytes`. So `crypto.getRandomValues` has never been used on Node.