@realityking 2025-08-03
I've updated the benchmark results.

I also realized that by keeping Node 18 compatibility (with the explicit webcrypto import) this PR has kinda lost its original purpose - increasing compatibility with non-node runtimes.

IMHO it'd still be good to merge this as it can be more easily stubbed but I'd also understand if you'd want to wait with this until dropping Node 18.