@broofa 2025-08-03
> I also realized that by keeping Node 18 compatibility (with the explicit webcrypto import) this PR has kinda lost its original purpose

I agree.  Unfortunately there doesn't seem to be a x-platform way of accessing `crypto` with node@18.  Not that I've found, at least.

This project supports current node "maintenance" version, plus one prior.  So we won't be dropping node@18 support until May 2026, when node@20 reaches end-of-life on node's release schedule.  Given that, I'm going to close this out for now, but I'll open an issue for addressing this after the uuid@12 release goes out.

Thanks for your time and effort on this.  Sorry we couldn't get it merged (yet).