@realityking 2025-08-04
Absolutely fair 👍 

I'll be back next year :)